+++ src/input/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
+export * from './exports';
 export * from './node';
 export * from './typings';