+++ src/input/utils.ts
@@ -1,5 +1,5 @@
 import * as ts from 'typescript';
-import { getLibRefName, getPropName, isBaseLib, getLib, getModule } from '../helpers';
+import { getLibRefName, getPropName, isBaseLib, getModule, getLibName } from '../helpers';
 import { DeclVisitorContext, TypeModelRef } from '../types';
 
 export function createBinding(context: DeclVisitorContext, lib: string | undefined, name: string) {
@@ -54,26 +54,34 @@ export function getSimpleRef(refName: string): TypeModelRef {
   };
 }
 
-export function getPackage(node: ts.Node, availableImports: Array<string>) {
+export function getPackage(node: ts.Node, global: boolean, imports: Record<string, Array<ts.Node>>) {
   const fn = node.getSourceFile()?.fileName;
-  const base = isBaseLib(fn);
-  const lib = getLib(fn, availableImports);
+  const base = isBaseLib(fn) || false;
 
-  if (base) {
-    return {
-      external: true,
-      moduleName: undefined,
-      base,
-      lib: undefined,
-      fn,
-    };
-  } else {
+  if (!base) {
+    const libName = getLibName(fn);
+    const [lib] = Object.keys(imports).filter(name => {
+      if (global) {
+        return name === libName;
+      } else {
+        return imports[name].includes(node);
+      }
+    });
+
     return {
       external: !!lib,
       moduleName: (lib && getModule(node)) || lib,
-      base: false,
+      base,
       lib,
       fn,
     };
   }
+
+  return {
+    external: true,
+    moduleName: undefined,
+    base,
+    lib: undefined,
+    fn,
+  };
 }